requirements.txt
django==5.1.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
